removed all tests from sln
